0. Modify the maximum upload file size. 1. Fixed the bug that the reading count is not displayed.

+++ news-service-article-html/src/main/java/com/aiolos/news/component/ArticleHtmlComponent.java
@@ -1,6 +1,7 @@
 package com.aiolos.news.component;
 
 import com.mongodb.client.gridfs.GridFSBucket;
+import org.apache.commons.lang3.StringUtils;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
@@ -28,6 +29,9 @@ public class ArticleHtmlComponent {
     }
 
     public Integer download(String articleId, String articleMongoId) {
+        if (StringUtils.isBlank(articleId) || articleId.equalsIgnoreCase("null")
+                || StringUtils.isBlank(articleMongoId) || articleMongoId.equalsIgnoreCase("null"))
+            return HttpStatus.INTERNAL_SERVER_ERROR.value();
         try {
             File file = new File(articlePath + File.separator + articleId + ".html");
             OutputStream outputStream = new FileOutputStream(file);
@@ -40,6 +44,8 @@ public class ArticleHtmlComponent {
     }
 
     public Integer delete(String articleId) {
+        if (StringUtils.isBlank(articleId) || articleId.equalsIgnoreCase("null"))
+            return HttpStatus.INTERNAL_SERVER_ERROR.value();
         File file = new File(articlePath + File.separator + articleId + ".html");
         boolean deleted = file.delete();
         return deleted ? HttpStatus.OK.value() : HttpStatus.INTERNAL_SERVER_ERROR.value();
